resgate de premiacao

+++ GGT/resources/views/premiacao/DiretorExecutivo/index.blade.php
@@ -10,7 +10,8 @@
             <table class="table table-hover">
                 <thead>
                     <tr>
-                        <th></th>         
+                        <th></th>
+                        <th scope="col">Status</th>
                         <th scope="col">Vagas</th>
                         <th scope="col">Titulo</th>
                         <th scope="col">Descrição</th>
@@ -19,21 +20,54 @@
                     </tr>
                 </thead>
                 <tbody>
+                          
+                       
                     @foreach($premios as $p)
+                        
+                        <?php 
+                          $disponivel;
+                          date_default_timezone_set('America/Sao_Paulo');
+                          $data = date('Y-m-d');
+
+                          if ($p->limite_vagas > 0) { 
+                             $disponivel = true;
+                          } else { 
+                            $disponivel = false ; 
+                          }
+                        ?>
+                       @if($disponivel)             
+                         <thead>
+                              <tr>
+                                  <th scope="row"><td >Premios disponiveis</td></th>
+                                  <td >{{$p->limite_vagas}}</td> 
+                                  <td >{{$p->titulo}}</td> 
+                                  <td ><p>{{$p->descricao}}</p></td>
+                                  <td >{{$p->valor}}</td>
+                                  <td >{{$p->data_limite}}</td>
+                                  <td>
+                                  <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#modalEditar" data-titulo="{{$p->titulo}}" data-id="{{$p->id}}"
+                                    data-descricao="{{$p->descricao}}" data-valor ="{{$p->valor}}" data-vagas="{{$p->limite_vagas}}" data-time="{{$p->data_limite}}" >Editar</button>
+                                  </td>
+                                <td><a href="/premio/delete/{{$p->id}}" class="btn btn-primary btn-fill">Excluir</a></td>
+                              </tr>
+                          </thead>
+                          @else
                         <thead>
-                            <tr>
-                                <th scope="row"><td >{{$p->limite_vagas}}</td> </th> 
+                              <tr>
+                                <th scope="row"><td >Premios indisponiveis</td></th>
+                                <td >{{$p->limite_vagas}}</td>
                                 <td >{{$p->titulo}}</td> 
                                 <td ><p>{{$p->descricao}}</p></td>
                                 <td >{{$p->valor}}</td>
                                 <td >{{$p->data_limite}}</td>
                                 <td>
-                                <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal" data-titulo="{{$p->titulo}}" data-id="{{$p->id}}"
+                                <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#modalEditar" data-titulo="{{$p->titulo}}" data-id="{{$p->id}}"
                                   data-descricao="{{$p->descricao}}" data-valor ="{{$p->valor}}" data-vagas="{{$p->limite_vagas}}" data-time="{{$p->data_limite}}" >Editar</button>
                                 </td>
                               <td><a href="/premio/delete/{{$p->id}}" class="btn btn-primary btn-fill">Excluir</a></td>
                             </tr>
-                        </thead>			    
+                        </thead>
+                        @endif
                     @endforeach
                 </tbody>
             </table>
@@ -54,31 +88,31 @@
       <div class="modal-body">
       
       
-         <form method="post" action="{{url('/premio/criar')}}">
+         <form method="post" action="{{url('/premio/criar')}}" onSubmit="validar()">
              <div class="form-group">
             	<label for="nome">Título</label>
-				<input type="text" name="titulo" class="form-control" id="titulo" value="">
+				<input type="text" name="titulo" class="form-control" id="titulo" value="" required>
         	</div>
 			<div class="form-group">
 				<label for="descricao">Descrição</label>
-            	<textarea class="form-control" rows="6" name="descricao" id="descricao" ></textarea>
+            	<textarea class="form-control" rows="6" name="descricao" id="descricao" required></textarea>
 			</div>
 			<div class="form-group">
 				<label for="recompensa">Valor</label>
-				<input type="number" name="valor" id="valor" class="form-control" min="0" value="">
+				<input type="number" name="valor" id="valor" class="form-control" min="0" value="" required>
 			</div>
 
 			<div class="form-group">
 				<label for="recompensa">Quantidade de Vagas</label>
-				<input type="number" name="qtdVagas" id="qtdVagas" class="form-control" min="1" value=" ">
+				<input type="number" name="qtdVagas" id="qtdVagas" class="form-control" min="1" value="" required>
 			</div>
 
 			<div class="form-group">
 				<label>Data Limite de inscrição</label>
-				<input type="date" name="data_expirar" id="data_expirar" class="form-control" placeholder="AAAA-MM-DD" value="">
+				<input type="date" name="data_expirar" id="data_expirar" class="form-control" placeholder="AAAA-MM-DD" value="" required>
 			</div>
 					
-			<button type="submit" class="btn btn-success btn-fill">Salvar</button>
+			<button type="submit" class="btn btn-success btn-fill">Salvar</button> <input type="submit">
 			<a href="" class="btn btn-danger btn-fill">Cancelar</a>
      
          </form>
@@ -92,7 +126,7 @@
 
 <!-- Modal update-->
 
-<div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="premio">
+<div class="modal fade" id="modalEditar" tabindex="-1" role="dialog" aria-labelledby="premio">
   <div class="modal-dialog" role="document">
     <div class="modal-content">
       <div class="modal-header">
@@ -134,6 +168,52 @@
     </div>
   </div>
 </div>
+<!-- Script modal -->
+  @section('scripts')
+      <script>
+          $('#modalEditar').on('show.bs.modal', function (event) {
+        var premio = $(event.relatedTarget)
+        var titulo = premio.data('titulo')
+        var descricao = premio.data('descricao')
+        var valor = premio.data('valor')
+        var data_expirar = premio.data('time')
+        var vagas = premio.data('vagas')
+        var id = premio.data('id')
+
+        var modal = $(this)
+        modal.find('.modal-body #titulo').val(titulo)
+        modal.find('.modal-body #descricao').val(descricao)
+        modal.find('.modal-body #valor').val(valor)
+        modal.find('.modal-body #qtdVagas').val(vagas)
+        modal.find('.modal-body #data_expirar').val(data_expirar)
+        modal.find('.modal-body #id').val(id)
+      })
+
+
+      function validar(){
+
+        now = new Date
+
+        data = document.getElementById("data_expirar").value.split("-");
+        
+        if(data[0]<now.getFullYear()){
+            alert("Data invalida");
+            return false;
+        }else{
+          if(data[1]<now.getMonth()){
+            alert("Data invalida");
+            return false;
+          }else{
+            if(data[3]<now.getDay()){
+              alert("Data invalida");
+              return false;
+            }
+          }
+        }
+      }
+
+      </script>
+  @endsection
 
 @endsection
